move the other new stuff too
